Convert tau ID bitmasks to a sequence of numbers
Even though we could use the bitmasks to decide if a given tau passes a working point or not, the bitmasks are not that great when trying to visualize their distributions. Instead of constantly converting the bitmasks to consecutive integers inside an analysis, we compute these branches beforehand. This save computing time and the selector classes remain readable as well.