Finalize CRAB submission scripts
When executing test/launchall.sh with an argument that points to the file containing a list of samples line-by-line, the script determines whether the sample is a data sample or not; then proceeds to submit the cfg files to CRAB accordin
gly. The script determines if the user has set their environment correctly before doing anything.