Generate the cmsRun cfg file by running cmsDriver.py
Instead of maintaining a static template from which the cmsRun cfg files is generated from, we use cmsDriver.py to generate individual config files. Only two types of config files need to be generated -- one for the data and the other for the MC CRAB jobs. The advantage with the cmsDriver method is that we don't have to guess the correct settings in the config files; they're automatically derived from the command line parameters supplied by the user.

In our case the command line parameters are specified in launchall.sh script. The static config file (which we still keep for now just to have a fallback solution) uses autoCond dictionary to get the correct global input tag for the cmsRun jobs. With the cmsDriver approach it's not possible to directly make use of the autoCond dictionary, so we read the dictionary in the launchall.sh script before running cmsDriver. The correct input tag is passed to the cmsDriver command.

Tested with DO_DRYRUN=true.