Hardcode global tags instead of reading the from autoCond dict
The recommended tag for data (94X_dataRun2_ReReco_EOY17_v2) was not present in the autoCond dictionary, although it is a valid tag with the latest CMSSW version (9_4_2) compatible with the nanoAOD fork. Hardcoding these values makes it more obvious which tag was used and would thus not rely on the CMSSW version.