Do not edit cmsDriver cfgs in-place but use a static one
Also provided command line options for the launchall script:
  -f the input dataset file (mandatory)
  -d for performing CRAB dryrun
  -D cfg file for data jobs
  -M cfg file for MC jobs

The only downside currently is that some of the settings (e.g. if a given sample is data or MC, global tags and JSON lumi file) are hardcoded at least twice: in launchall.sh and the alternative static config files (in our case it's the cfg file to run jetSubstructure jobs). I don't think it's possible to propagate such settings to these static config files via environment variables as we do for the CRAB cfgs.