Solve the mass & charge problem for exotic particles once and for all
Previously we pulled the masses and charges of certain exotic generator level particles from https://github.com/alberto-sanchez/EvtGen/blob/master/evt.pdl but this turned out to be short-sighted for a couple of reasons. Firstly, the masses were hardcoded for certain selection of generator-level particles. However, this selection needed to be updated every time the code threw an error because we forgot to include yet another particle to the selection. We could parse the table used in the previous revision of the repository, but then again we really don't need to know this kind of information. If the low-mass particles are not defined in ROOT's PDG id table, it's more than likely that we don't need this information at all. So we set the masses and charges of such particles to 0. If it becomes a necessesity to know these masses, a proper solution involves parsing an officially produced table that contains the masses and charges of exotic particles. But for now, let's go with something that definitely works for us at this time.