Add b-tagging variable to leptons for fakeable selection
We want to migrate to DeepCSV score, but the b-tagging WPs used in fakeable electron selection are not known, yet. For the purpose of sychronization, we'll add a branch which tells us if the jet nearest to a given fakeable electron is below 0.3 (0.8484) if the electron fails (passes) the MVA > 0.9 cut or not. Notice that these b-tagging WPs are for the old CSV discriminator. Once we have corresponding WPs for the DeepCSV discriminator, the changes in this commit can be dropped.