Determine histogram binning after reading the 1st event
The problem is that some MC samples have different number of LHE pdf weights, which means that we have to read the 1st event in before we initialize the histograms. This has two side-effects compared to previous implementation:
- there will be only Count histogram in data Ntuples;
- if the Ntuple contains no events, no histograms will be created.
Previous implementation at least initialized the histograms in both cases (but left them empty). This is probably a better solution than propagating the number of pdf weights via constructor or setting the number of bins to a high and removing any checks regarding the mismatch b/w the expected and true nof pdf weights.